feat: models

+++ internal/model/user/user.go
@@ -7,5 +7,5 @@ type User struct {
 	Username string    `json:"username,omitempty"`
 	Email    string    `json:"email,omitempty"`
 	Password string    `json:"password,omitempty"`
-	Balance  float64   `json:"balance,omitempty"`
+	Balance  float32   `json:"balance,omitempty"`
 }
